requirements.txt
openai==1.30.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
requests==2.31.0
